requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
psycopg2-binary==2.9.10
redis==6.4.0
rq==2.5.0
requests==2.32.5
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
pytest==8.4.2
